Make template mask immutable
Some functions (e.g. parcellation.plot_surface) pass the template
file name to another program, which reads it from disk.

If the user edits p.template_mask in memory e.g. if they want to
add a voxel to the parcellation, then those functions will return
incorrect/misleading results because the changes to p.template_mask
are not propagated to the .nii file. A safe workflow is for users to
first modify the template mask on disk, and never modify it within
the parcellation object