Fixed loading parcellation objects from disk
Note that because the template filename is stored as a full path
it could be different if the parcellation object is shared. So
at the moment, parcellation objects should not be shared across
users, parcellations should be stored on disk and then this class
instantiated when doing computations. There doesn't appear to be
an easy fix for this, one option would be to store the header in
the parcellation object as well and thus do away with the need for a
template nii file entirely (after creating the parcellation object).
However, this will rely on having a better way to write nii files
with particular headers without using fslcpgeom