Voxels with no parcels get zero value rather than NaN
In some parcellations, there are voxels in the standard brain mask
that are not assigned to any parcel. When parcel values are
projected onto voxel space, what value should these be assigned?
Really, these ought to be NaN. However, in the output NIFTI file,
the image usually has a value of zero outside the brain anyway. The
other thing is that workbench returns poor results with the Fieldtrip
HCP mask if NaNs are present outside the brain surface (which is the
case due to the expansion of the mask compared to the brain) and
fsleyes displays the colour of zero if the value is NaN and there is
no way to hide it. So it's more convenient to set those voxels to
zero. This also matches the fact that matrix2vols assigns a value of
zero outside the brain already, so there is some precedent for this.